requirements.txt
# Brasileirão Analytics — Dependências do Projeto
# Gerado para Python 3.10+

# --- Web Scraping (Camada Bronze) ---
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0          # Parser HTML/XML de alta performance para BeautifulSoup

# --- Processamento de Dados ---
pandas>=2.2.0
numpy>=1.26.0

# --- Banco de Dados (Camada Ouro / Load) ---
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0        # Driver PostgreSQL (v3 — resolve bug de locale no Windows)

# --- API REST ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# --- Utilitários ---
python-dotenv>=1.0.0     # Leitura do .env
